requirements.txt
numpy==2.3.5
pandas==2.3.3
Pillow==12.0.0
scikit_learn==1.7.2
torch==2.5.0
torchvision==0.20.0
tqdm==4.67.1
